Add test to test what is happening when an error is raised in conditions
fix: Exception in haskell condition gracefully shutdown solver

In the event of an exception in the Haskell condition code
(`EventConditionsHaskell`), the RTS was crashing the program.

Instead, we force the evaluation (i.e. the exception may be hidden in a
lazy computation) with `force` / `evaluate`, and return a non 0 value so
sundial knows that the solving failed in the root finding function,
according to https://sundials.readthedocs.io/en/latest/cvode/Usage/index.html#rootfinding-function

Tests are added to ensure that it correctly fails.