feat: cooperative solver stop in case of async exception
If an async exception (such as a timeout, are thread cancel) was
happening when the solver was the inline-c block, nothing was happening
and the exception was ignored until the inline-c block terminates, which
could take a long time if the solver was working on an hard task.

This commit improves the situation:

- In the haskell thread, we wait for exception. If any happen, we set
  `*ptrStop = 1`. This value is visible from the FFI call.
- In the solver (i.e. the FFI call), we test `*ptrStop` after each
  timepoints and stop the solver if required, after the necessary
  cleanups.

A few implementation details:
- if the solver is stopped by an exception, the `FFI` call will
be terminated thanks to the `ptrStop` logic AND the exception will be
raised again once terminated. So it is not possible to stop the solver
AND use the result by mistake.
- The `ptrStop` cannot just be read at each loop without proper
  synchronisation. In practice, the loop is complex enough so the
  compiler had really few chances to not generate the read for
  `*ptrStop`, but this is not guaranted. Using the `__atomic_load`
  ensures that proper barrier for compiler and CPU are set.