feat: propagate exception in cond/rhs to caller
Before this commit, an exception in the Haskell callback for condition
and ode rhs had different results:

- In a condition, the exception was ignored and sundials was forced to
  fail, resulting in an "opaque" error for the calling user.
- In an rhs, the exception was caught by the RTS which was then crashing
  the program.

This commit changes the following: now the exception is caught by the
handler code, sundial is forced to terminate, and the exception is
throw again by the call site.

This lead to a more robust solving experience: no crash of the complete
program AND a precise exception.

Note also that the exception code for condition was changed a bit and
does not involve `force` nor `evaluate` (which is implicitly done by
the `poke` operation) and it should hence be a bit faster.

Tests were updated to detect the new exception instead of detecting a
mute solver error.