fix: vector coerce is stricter
`coerce` was used to convert a `StorableVector a` to `StorableVector b`
as long as `Coercible a b`.

This is no longer possible because there is no guarantee that `a` and
`b` share the same `Storable` instances. We need to use `unsafeCoerce`
here unfortunately.

In some condition, it would be possible to use `unsafeCoerceVector`
which is a bit safer, `unsafeCoerce :: a -> b` when `unsafeCoerceVector
:: Coerible a b => Vector a -> Vector b`, but tweaking the code manually
for each occurrence is too painful.