feat: add odeOnTimePoint, callback for timepoint solving
If `odeOnTimePoint` is not `Nothing`, a callback function can be
registered. This function is called each time a timepoint is registered,
3 time in each C integration code:

- Once after initial condition are set
- For each defined timepoint
- For each extra timepoint generated by events

It should have no (or negligible) impact on performance.

The callback is really simple for now, it only accept the current
timepoint index as argument.

Having this callback called when the solver actually store the value
will give the caller an opportunity to compute and store other metrics,
such as ram usage, real time, ...

In the future, we would like to:

- Evolve the solving api so instead of being a callback, the solver
  behaves as a `Stream` of timepoint. This would allows even more
  control on the metrics or behavior and maybe nice feature such as
  pause / save of the solving context and resume later.

Regarding exception, if an exception happen in the callback, it will be
caught by the `saveExceptionContext` logic, but solving will continue.
That's actually imperfect and should be improved later.