refactor: update golden tests
Only ARKode golden test had changes, on really small and limited values.
Maybe this needs to be investigated, but seems related to a small change
in sundials.

The `canonical/Robertson.json` result is surprising, especially on the
number of steps and evaluation listed in debug, but results are kinda
alike.