refactor: complete rewrite of cvode with haskell instead of inline-c
Message to be detailled, but all the inline C logic was rewritten in
pure haskell.

In order to do so, I had to:

- foreign import most required C functions.
- Use a `StateT` to track the loop iterative mutable index.

A lot of refactoring could happen later in order to clean / simplify /
extend the API, but the idea was to get as quickly as possible close to
the original implementation.

A few semantic changes however:

- The code won't leak resources anymore in most contexts, especially, in
  case of exception or early exit, the sundial context / memory /
  vectors are correctly freed. Restoring the previous / bugged behavior
  was actually too painful.