refactor: complete rewrite of cvode with haskell instead of inline-c
The inline C logic was rewritten in pure haskell in cvode and arkode.

This rewrite was done in order to be "easy" to review and confirm that
nothing was changed, important refactoring / cleaning may follow:

- the "main iterative loop" was replaced by a `StateT` to track the loop
  variables.
- All the required functions were imported with `safe` foreign call.

A few semantic changes however:

- The code won't leak resources anymore in most contexts, especially, in
  case of exception or early exit, the sundial context / memory /
  vectors are correctly freed. Restoring the previous / bugged behavior
  was actually too painful.
- The `ptrStop` logic is no longer required, the RTS will be able to
  rethrow async exception everytime it comes back into the Haskell loop,
  so the only blocking event should be a really long / infinite loop in
  `CVode` / `ARKodeEvolve`.

This should have no important impact on performance (most of the "cost"
of computation is in the `CVode` and `ARKodeEvolve` function), it should
however allocate a bit more Haskell values during the loop (should be
minimal). On the other hand, this refactor will allow more correctness
work in the future (we already saw correctness improvement with memory
behavior). It should be simpler to extend by other developers not at
ease with C, Haskell and foreign call logic. `inline-c` was difficult to
work with, especially in ghci or HLS, that's not a problem anymore.
Finally, it will allow change in the algorithm, such as turning the main
loop into a `Stream (Of (CurrentTime, [TimeMeasures], [Diagnostic]))`
(e.g.) allowing more control on the complete logic, such as introducing
stop, or introducing more diagnostics.