feat: Introduce IDA
The golden test files are copied from CVode results, because it should
be the same.

`IDA.hs` is copied and adapted from `CVode.hs`.

A lot of change happened in the `Common.hs` and `Sundials.hs` to fit the
new API.

For the user, the `odeRhs` from `OdeProblem` was replaced by another
structure which can stores the required values for `IDA` solving.

Solving with `IDA` a standard "ode rhs" problem is doable (that's
actually what tested by most of the default unit test) and this gives us
a lot of confidence that the implementation is correct.

However, solving a rhs problem using arkode or cvode will simply fail.