feat: completly disable sundials logging
Sundials logging can spam the output with millions of message and the
logs messages can only be returned to files.

We disable them for now. Maybe in the future we'll affect a unique file
per solving process and then re-read that file in order to output a
subset of the relevant logs.

Note: code is duplicated between ARKode and CVode, another WIP MR is
currently trying to factorize this. This is an orthogonal work, so we
don't bother too much right now.