refactor: remove the now useless ptrStop semantic
This was done in order to allow the solver to stop in case of exception
on specific steps of the solver loop.

However now the solver is mostly written in Haskell, so it can stop on
most steps (actually more than before) and the ptrStop semantic was not
used anymore.