feat: disable the "canonical" tests
These canonicals tests are more pain that solution. When creating the
tests for new methods (such as IDA), we can take great care to check
that the references are not too much different. Then the test is fixed
and will live its life.

The canonical tests introduces surprise in the test process, because it
is possible that completely different methods (such as IDA) generates
small variations, and because all the methods are writing on the same
file, it leads to confusion when the test fails for some method, but not
other, and once `accept`ed, the opposites happen.