fix(ida): correct values at t=0
This fix is surprising and I don't fully understand it.

I observed that the value at t=0 was not taking into account the
correction introduced by `IDACalcIC`, we need to get
`IDAGetConsistentIC` to get the correct value (and override `y` and `yp`
vector).

Doing this fixs the newly introduced test where we can observe that the
fixed value appears correctly in the output vector at t=0.

It is unclear if the fixed value was already taken into account during
solving, because when I introduced `IDACalcIC`, I got result (e.g.
solving was able to get better results).