fix(ida): do not crash when event happen at t=0
See point 2 in https://github.com/novainsilico/hmatrix-sundials/issues/38

During an event, we need to recompute the algebraic variables using
`IDACalcIC`. This function requires a "next time" value for its internal
computation and we need to ensure that this value is in the future. This
was done using `next_time * 1.1`, which is in the future for all values,
except `0`.

This commit fixs this by adding a special case, if `next_time == 0`, we
use `1`. This is completely arbitrary, but at least it fix the crash,
we'll evaluate later if a better solution can be designed.

We also added a unit test for this.