feat(ida): handle discontinuity in algebraic rules
A discontinuity in an algebraic rule would lead to a crash of the
solver. We already handled this case when an event would change a
variable involved in an algebraic rule, but we also need to fix this in
the context were the discontinuity happen without any event.

The solution is that we jump over the discontinuity, recompute a valid
state (with `IDACalcIC`) and restart the solver. Extra care need to be
taken when the discontinuity is close to a stop time.

Added 3 testes:

- One when the discontinuity happen far away from a stop time.
- Two when the discontinuity happen at the stop time (inclusive or
  exclusive, e.g. < stop_time or <= stop_time). Note that in the case of
  a discontinuity happening exactly at the stop time, the solver records
  the previous value at the time of the recording, which is coherent.