fix: jump over discontinuities
When IDA find a discontinuity, the solver will fail with different
possible error code. Either because it cannot solve the constraints, or
because it backtracked too much time without any progress. This also
happen when the current time is large compared to the minimum step
size.

The solution is, as what we've done for event, we just try to jump over
the discontinuity and continue. This is imperfect, and at risk of
missing some important features (such as event), but in the current
context, it feels better than a solver failure.