ida: do not double RecalcIC after event
When the event handler is called, it is responsible to call the callback
which would recalculate the initial condition in the event of a change
of algebraic equations.

However, we were also doing that a second time unconditionally after an
event handling.

This is a performance waste considering that the event handler may had
always done that or maybe that the event handler knows that this is not
required.

We removed this call and now require that the event handler do the call
when required.