feat: only reinit when required
When reinit is False, we do not reinit, allowing for some event to not force the timestep to small size, giving a boost of performance.

This can happen in two context:

a) The event does not do any change (for example, it was removed by some condition)
b) The event does a change on a value which have no impact on the ode system, such as a counter. In this context, the solver can continue without being impacted.